consolidated aya containers to single folder moved aya hero styling into config
