+++ src/components/HeroPanel/HeroPanel.js
@@ -10,6 +10,7 @@ export default class HeroPanel extends Component {
     title: PropTypes.string,
     children: PropTypes.node,
     style: PropTypes.string,
+    styles: PropTypes.object,
     type: PropTypes.string,
     isEmpty: PropTypes.bool,
     hasBreadcrumb: PropTypes.bool,
@@ -19,6 +20,7 @@ export default class HeroPanel extends Component {
   render() {
     require('./HeroPanel.scss');
     const {title, subTitle, image, isEmpty, children, hasBreadcrumb} = this.props;
+    const styles = this.props.styles || {};
     const type = get( this.props, 'type', 'site-heading');
     let className = 'hero-panel';
     let style = this.props.style;
@@ -36,14 +38,14 @@ export default class HeroPanel extends Component {
               <div className="image-wrapper-holder" />
               <div className="image-wrapper-overlay" />
             </div>
-            <div className="hero-panel-type">
+            <div className={`${styles.defaultBg} hero-panel-type`}>
               <div className="hero-panel-type-lining">
                 <div className={type}>
                     <div className="col-lg-8 col-lg-offset-2 col-md-10 col-md-offset-1">
 
                       <If condition={!isUndefined(title)}>
                         <h1>{title}</h1>
-                        <hr className="small" />
+                        <hr className={`${styles.hr} small`} />
                       </If>
                       <div className="subheading">
                         <If condition={subTitle}>
