+++ src/containers/Aya/heroPanel.scss
@@ -1,0 +1,29 @@
+@import 'src/theme/variables';
+
+.hr {
+    border-top: 1px solid #fff;
+    height: 30px;
+    margin-left: 30px;
+    margin-right: 30px;
+    overflow: visible;
+    &:before,
+    &:after {
+         background-image: url('./images/ayaLeaf-left.png');
+         height: 21px;
+         background-repeat: no-repeat;
+         content: ' ';
+         width: 30px;
+         position: relative;
+         bottom: 11px;
+         float: left;
+         margin-left: -30px;
+    }
+    &:after {
+         background-image: url('./images/ayaLeaf-right.png');
+         float: right;
+         margin-right: -30px;
+    }
+}
+.defaultBg{
+    background-color: $aya-green;
+}
