+++ src/containers/Aya/images/ayaLeaf-right.png
