+++ src/containers/index.js
@@ -2,7 +2,7 @@ export App from './App/App';
 export DefaultApp from './App/Default';
 export AyaApp from './App/Aya';
 export Home from './Home/Home';
-export AyaHome from './Home/AyaHome';
+export AyaHome from './Aya/Home';
 export RecipeList from './Recipes/RecipeList/RecipeList';
 export ViewRecipe from './Recipes/ViewRecipe/ViewRecipe';
 export UpdateRecipe from './Recipes/UpdateRecipe/UpdateRecipe';
@@ -37,5 +37,5 @@ export UpdateLink from './Links/UpdateLink/UpdateLink';
 export DeleteLink from './Links/DeleteLink/DeleteLink';
 export LinkList from './Links/LinkList/LinkList';
 export JonProfile from './Profile/JonReading';
-export AyaArticles from './Profile/Aya';
+export AyaArticles from './Aya/Articles';
 export AyaProducts from './Aya/AyaProducts/AyaProducts';
